chore: moved unit tests into /unit folder
