+++ tools/erc-repository-indexer/erc-contract-indexer/tests/unit/utils/constants.ts
